Refactor cease & revoke to standard submit method
Previously the deregistration page object had a `revoke()` and `cease()` method, which did the same thing only they selected a different value in the page's reason drop down.

This change refactors them into a single `submit()` method, with the reason passed in as an argument. This also brings the page inline with the standard pattern of each page object featuring a `submit()` which handles completing and submitting said page.