Refactor admin nav bar functions to Section
This change adds a [SitePrism section](https://github.com/natritmeyer/site_prism#sections) which contains all the logic needed for selecting the elements of the nav bar in the back office system.

A **SitePrism** section

> allows you to model sections of a page that appear on multiple pages or that appear a number of times on a page separately from Pages.

So in our case once a user is signed into the back office, the nav bar appears on every page. By creating a section object we can interact with the nav bar from each page, but maintain all the logic in one place.