Fix broken registration journeys v2!
Having fixed the journeys in 05f9102 we then also deployed the latest version of WEX, which then broke the journeys again!

This was because of a change in behaviour to the choose exemptions page. Now clicking submit on moves you to the next page if you are on the last tab (storing waste), else it just moves you to the next tab.

Because of this any step which selected an exemption which wasn't on the last tab, or which selected and **S1** followed by a **D2** would put the user in a tab that wasn't the last. Clicking submit then would only move the user to the next tab, whereas the tests assume you have moved to the next page.

This change fixes these issue.