Fix incorrect value used in assertion
This fixes one of the tests in the `backoffice_site_address` feature, which asserted that if a site address was selected using the postcode CV35 9ES (and the last address selected) then the service would determine the NGR was SP2518455455. In fact it should be SP2604755720.

This change corrects the test.