Fix BO site address with NGR feature examples
https://eaflood.atlassian.net/browse/WAS-1149

We initially thought that a fix we had put in place for this issue (enter a site using an NGR and the EA admin area doesn't get updated) wasn't working because scenario *Registration by a NCCC user using a National Grid reference (NGR) for site address - area is added to registration* was failing.

However upon investigation we found it was actually bum data in the test itself; the area name was wrong for the NGR used.

This simply corrects the scenario example data.