Move page object sections into '/sections' folder
This change moves the all classes in the `/page_objects` folder that inherit from `SitePrism::Section` into their own `/page_objects/sections` folder.

This will help improve maintenance (the fact we use sections and where to find them is now obvious), but is also part of sorting issue #9 . The issue is when you have a page object called `RegistrationPage`, and you want to use something in the same folder called `SiteDetailsSection`, you can't because of the order in which they are loaded.

Placing them into a separate folder forces us to use `require_relative`, which in turn resolves the issue of the order the classes are loaded (it no longer matters).

Finally because of the way things are loaded into a Cucumber environment, we found placing multiple `require_relative`s to the same class was causing the following message to appear in the output; `warning: already initialized constant SiteDetailsSection::SELECTOR`. Memoizing was felt to be a perfectly safe solution, as we know this is just a result of loading the same class multiple times. See http://stackoverflow.com/a/23552366/6117745