Rename LocationDetailsSection 2 SiteDetailsSection
https://github.com/DEFRA/waste-exemptions-acceptance-tests/issues/9

This fixes issue #9, which was that when attempting to refer to a class called `SiteDetailsSection` from one called `RegistrationPage`, because of the order the files are loaded `RegistrationPage` errored as it did not recognise `SiteDetailsSection`. Naming it `LocatioDetailsSection` forced it to come before the page in the load order and hence the error went away.

However this is not ideal, and we really want the name of the class to represent the section of the page it covers. We initially thought we'd have to restructure the whole project, however after some testing a simple `require_relative` to the section class solved the problem.

Rather than make an exception for this one case, we made `require_relative` standard for any page referencing a section in 387b1c8, so all that's actually changed in this commit renaming of the file and class, and updating all references to it.