Update conf. letter export ft. to chk for content
https://eaflood.atlassian.net/browse/WAS-1152

We recently put the confirmation letter generation feature live, however when testing the bulk letter export we spotted an issue only after it had ben released.

Assume today is 3 March 2017. If I was to add a registration (and only one) and then go to the bulk letter export, and select 2017-03-03 to 2017-03-03 it would say there were no registrations to export.

Selecting 2017-03-02 to 2017-03-03 doesn't help either. It would therefore seem that the bulk letter export only works for registrations created prior to the current date.

So we have updated the `Back office user exports confirmation letters` scenario in `confirmation_letters_export.feature` with an additional step that will check that the filename is set to something other than 'n/a', as this will confirm that the registration was included in the export. However we have set it to pending for now as this is a known issue we want to fix, but we don't want it causing the acceptance tests when run in our CI environments to fail.