Fix failing register as a partnership scenario (#56)
Scenario 'Registration by a partnership' in `register_new_exemption.feature` was failing. After investigation it was identified that a new address is now returned by OS Places when using the postcode BS1 5AH. We use this address of the first partner to pre-populate the correspondance address, however `premises` is a mandatory field on that screen and this new address was missing it.

So this change simply passes in a set premises name to use when completing the correspondance details as part of the scenario.