Add workaround for tests failing in pre-prod env (#59)
As of writing it was 8 months since we had run the acceptance tests in our environments. When attempting to deploy an update recently the acceptance tests would not connect to the chromedriver.

We were able to resolve that by putting pack the version of the chromer-driver-helper gem used by the project, but then found that the tests would fail on any page where the submit button was not visible in the window i.e. you would have to scroll to reach it.

We could not explain why this has happened (the content of the pages has not changed and only minor updates to the projects dependencies has taken place) but were able to get the tests passing if we resized the window such that the continue button is always visible.

This change includes the functionality for that workaround.