Fix back office user credentials
Recent runs of the acceptance tests have been failing, and upon investigation it looks like this is because the previous credentials no longer work.

This change brings in the new way of holding credentials i.e. we no longer commit them to source control and behind the scenes we've updated the credentials.