Unpin chromedriver-helper version
Previously we had to pin it to an older version in order to match the versions installed in our Jenkins CI environment.

However they have now been updated so the tests no longer run. This change removes the pinning and updates the version so we match what is now in our Jenkins CI environment.