Update project following HO from @timstone
With the departure of @timstone from the team @cruikshanks is picking up QA responsibilities in the interim.

These changes are as a result of starting to use the project. Specifically when running tests against an environment I think it will be easier to do if the majority of configuration is already setup, rather than having a single file you have to more heavily modify.

So this creates a config for each non-production environment, and removes `.config.yml` so you users of the project have something which won't be committed and which they can specifically use for ad-hoc testing.

Also it moves all the config into a new folder so we can better understand which config files relate to which testing scenario.