Implement defra-ruby-style
This implements [Defra Ruby Style](https://github.com/DEFRA/defra-ruby-style) which is a Defra gem used for managing our rubocop rules and configuration across projects.

Along with implementing the gem this change resolves issues found when running the new rules. Finally we made some tweaks to the Travis CI config to bring the project more inline with other and to reduce the complexity in the `Rakefile`.