Tidy up of the config files (#129)
Prior to this change the config files were dot files, which meant by default they would be hidden on nix based systems.

There is no need for them to be hidden, in fact they should be visible. Also they frequently used capital letters which is seen as non-conventional as it requires extra key strokes to access when working in the terminal.

So this change updates the file names to apply a consistent pattern and ensure the files are visible.

We have also moved the ci config file into the config folder, so we are consistent with the location of all config in the project.

Along with the file name changes we have also tidied up the `Rakefile` so that what it offers matches the config we have. Again we have removed capitals from the rake task names to match standard conventions.