Update to match WEX technical alignment version
When looking at the existing coverage of the acceptance tests I found that it was not in alignment with the technical alignment work we have been doing in WEX.

This meant there was a lot of functionality for the old version of the service which as yet is not in the new TA version. Also functionality that is in the TA version had no coverage from the acceptance tests.

Hence this change which aims to bring the acceptance tests in line with the version of WEX we are currently building and intend to deploy.