Update to match WEX technical alignment version (#130)
When looking at the existing coverage of the acceptance tests I found that it was not in alignment with the technical alignment work we have been doing in WEX.

This meant there was a lot of functionality for the old version of the service which as yet is not in the new TA version. Also functionality that is in the Technical Alignment version had no coverage from the acceptance tests.

Along with this change were some more fundamental ones to how things are done. The list below provides some details

- Centralise where data is stored into a single helper
- Move all registration creation logic into a single place as a helper to remove duplication
- Use more dynamically generated data using the [Faker gem](https://github.com/stympy/faker)
- Add a mechanism for pre-generating data using the `@data` tag. The `before` hook will see this and for the first scenario it will generate the data. Subsequent calls will then be ignored. This saves time in the tests because in some cases the same data, even if amended by a prior scenario can be reused
- Add a mechanism for storing state between scenarios; `@my_var` instances will hold state between steps when running a scenario. However to hold state between scenarios it needs to go into a global `$my_var`
- Add the `features/support/world.rb` object. This is more featured than the previous `features/page_objects/app.rb` and is intended to represent everything in the world being tested including pages, retained state, and access to our custom config
- General tidy up and re-organisation of where app and page object files go