Link quke dependency to github instead of rubygems
The master branch of Quke which is the framework we use for testing contains some fixes we want to take advantage of in the acceptance tests. However it's not yet ready to cut a release so the version pushed to ruby gems is behind.

This changes updates the project to track Quke from the github repo rather than rubygems.

It's not expected to be a long term thing.