Add cover for resume a registration
https://eaflood.atlassian.net/browse/RUBY-199

If an external user starts a registration but isn't able to complete it, and they call NCCC for help, NCCC staff should be able to pick it up and continue working on it.

We have added the functionality for resuming a registration from the back office so this change provides acceptance test cover of it.