Fix BO create reg using front office (#136)
We spotted that there was an inconsistency with the back office create registration feature. After clicking the link in the back office to start a new registration it was then switching to the front office to actually complete the journey.

We had simply forgotten to pass in the param to tell the `add_submitted_registration()` method not to load the homepage when it starts.