Fix typo and update defaults in config files
We spotted that there was a typo in the name of the parallel section in all 4 environment `.config.yml` files.

This change fixes that, but also updates some of the defaults to match current preferences

- use headless is now `false`
- print_progress is now `false`
- enable parallel is now `false`

The project is being picked up by a new QA (👋 @andrewhick ) and as such running with these defaults will give a better sense of what is going on.