Add data exports check to user roles feature
https://eaflood.atlassian.net/browse/RUBY-68

This change adds steps that confirm all user roles have access to the data exports page and that when they click through to it the page loads.

It's difficult to go beyond this because the content is very dependent on what data was created when in the database, but hopefully this is sufficient to at least confirm the page is up and the permissions are correct.