Fix last email urls for AWS testing
https://eaflood.atlassian.net/browse/RUBY-1062

As part of ensuring our next update to production we spotted that WEX acceptance tests were no longer running 'green'. After some investigation it seemed the urls `/email/last-email` were no longer available.

Some more poking found that you simply need to go to `/last-email` to access them. This may be down to the install of Ruby 2.7.1 in the environment. But TBH we don't know. We just know `/last-email` works!