Reverting config change
It seems what URL is used depends on the version deployed. I currently believe that the changes we did to support the upgrade to Rails 6 has sorted the inheritence and means the feature is available on `/email/last-email` in DEV and TEST. And that is where we expect it.

PRE-PROD currently has a pre-rails6 version deployed so my belief is that is effected by the issue @cintamani found in the mocks. We found that how a controller was built at runtime was not consistent. This would explain why `/email/last-email` has worked, then suddenly you need to go to just `/last-email`.