Use /last-email for email testing everywhere (#211)
Currently, the tests will look in Mailcatcher for emails when running locally and at `/last-email` when running against a remote AWS environment.

This initially masked a problem finally exposed in [PR #210](https://github.com/DEFRA/waste-exemptions-acceptance-tests/pull/210). So to avoid any confusion or issues in the future this update ensures we grab email content from `/last-email` whatever the environment.